requirements.txt
Jinja2
python-dateutil
paramiko
scp